+++ clients/derivatives_trading_coin_futures/src/binance_sdk_derivatives_trading_coin_futures/websocket_streams/websocket_streams.py
@@ -33,7 +33,6 @@ class DerivativesTradingCoinFuturesWebSocketStreams(WebSocketStreamBase):
         self,
         configuration: ConfigurationWebSocketStreams,
     ) -> None:
-
         super().__init__(configuration)
         self.configuration = configuration