+++ clients/spot/src/binance_sdk_spot/websocket_streams/websocket_streams.py
@@ -41,7 +41,6 @@ class SpotWebSocketStreams(WebSocketStreamBase):
         self,
         configuration: ConfigurationWebSocketStreams,
     ) -> None:
-
         super().__init__(configuration)
         self.configuration = configuration