+++ clients/alpha/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.0 - 2026-01-29
+
+### Changed (1)
+
+- Updated `binance-common` library to version `3.5.0`
+
 ## 1.0.0 - 2026-01-23
 
 - Initial release