+++ clients/derivatives_trading_portfolio_margin_pro/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 5.4.0 - 2026-01-29
+
+### Changed (1)
+
+- Updated `binance-common` library to version `3.5.0`
+
 ## 5.3.0 - 2026-01-23
 
 ### Changed (1)